app/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
f==0.0.1
Flask==2.2.3
Flask-HTTPAuth==4.8.0
Flask-PyMongo==2.3.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
PyYAML==6.0
requests==2.29.0
six==1.16.0
urllib3==1.26.15
Werkzeug==2.3.0

